requirements.txt
asgiref==3.5.2; python_version >= '3.7'
django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0; python_version >= '3.6'
sqlparse==0.4.3; python_version >= '3.5'
